tbf/pyproject.toml
[tool.poetry]
name = "tbf"
version = "0.1.0"
description = ""
authors = [""]
license = "MIT"

[tool.poetry.dependencies]
python = "3.9.6"
Django = "^4.0.2"
uWSGI = "^2.0.20"
uwsgitop = "^0.11"
psycopg2 = "^2.9.3"
tqdm = "^4.62.3"
Pillow = "^9.0.1"
moviepy = "^1.0.3"
h2o = "3.36.0.2"
pandas = "^1.4.0"
jupyterlab = "^3.2.9"
kaleido = "0.2.1"
dtale = "^1.61.1"
pandasgui = "^0.2.13"
publicsuffix2 = "^2.20191221"
connectorx = "^0.2.4"
ipywidgets = "^7.7.0"
anytree = "^2.8.0"
chefboost = "^0.0.17"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
